move list components outside view
also removed unused `list-item` component
