+++ app/components/index.ts
@@ -9,8 +9,8 @@ export * from "./route-card/route-card"
 export * from "./route-details-header/route-details-header"
 export * from "./change-direction-button/change-direction-button"
 export * from "./text-input/text-input"
-export * from "./list-item/list-item"
 export * from "./route-details/route-details"
 export * from "./star-icon/star-icon"
 export * from "./favorite-routes/favorite-routes"
 export * from "./date-picker-modal"
+export * from "./list"
