+++ app/components/list/index.ts
@@ -1,0 +1,2 @@
+export * from "./list"
+export * from "./list-item"
